data_generation/requirements.txt
Faker==13.3.4
kafka-python==2.0.2
psycopg2-binary